update loans logic

+++ app/(app)/expenses/loans/LoansTable.tsx
@@ -39,39 +39,52 @@ import { ILoan } from "@/types/ILoan";
 import { InstallmentInterface } from "./utils/installments";
 import apiClient from "@/config/apiClient";
 import { LoansInstallments } from "./LoansInstallments";
+import { DialogForm } from "@/components/dialogs/DialogForm";
+import { LoanForm } from "./LoanForm";
+import {
+  getLoanStatusColor,
+  getLoanStatusText,
+  LoanStatusColor,
+  LoanStatusText,
+} from "./utils";
 
 export function LoansTable() {
   const [sorting, setSorting] = useState<SortingState>([]);
-  const { loans } = useContext(LoansContext);
-
+  const { loans, deleteLoan } = useContext(LoansContext);
+  const [openForm, setOpenForm] = useState(false);
+  const [selectedLoanId, setSelectedLoanId] = useState<string | null>(null);
   const [installments, setInstallments] = useState<InstallmentInterface[]>([]);
+  const [deletingLoans, setDeletingLoans] = useState<string[]>([]);
 
-  const deleteLoan = async (loanId: string) => {
-    try {
-      await apiClient.delete(`/loans/${loanId}`);
-      // Optionally, you can refresh the loans list after deletion
-      // fetchLoans();
-    } catch (error) {
-      console.error("Error deleting loan:", error);
-    }
-  };
-
-  const handleDelete = (loanId: string) => {
+  const handleDelete = async (loanId: string) => {
     if (confirm("¿Estás seguro de que deseas eliminar este préstamo?")) {
-      deleteLoan(loanId);
+      if (deletingLoans.includes(loanId)) return;
+
+      setDeletingLoans((prev) => [...prev, loanId]);
+      try {
+        const res = await apiClient.delete(`/loans/${loanId}`);
+        if (res.status === 200) {
+          console.log("Loan deleted successfully");
+          deleteLoan(loanId);
+          setDeletingLoans((prev) => prev.filter((id) => id !== loanId));
+        }
+      } catch (error) {
+        console.error("Error deleting loan:", error);
+        setDeletingLoans((prev) => prev.filter((id) => id !== loanId));
+      }
     }
   };
 
   const fetchInstallments = async (loanId: string) => {
     try {
       const response = await apiClient.get(`/loans/installments/${loanId}`);
-      const data = response.data;
-      console.log(data);
-      setInstallments(data);
+      console.log("Installments fetched successfully", response.data);
+      setInstallments(response.data);
     } catch (error) {
       console.error("Error fetching installments:", error);
     }
   };
+
   const columns: ColumnDef<ILoan>[] = [
     {
       accessorKey: "createdAt",
@@ -96,7 +109,7 @@ export function LoansTable() {
           onClick={() => column.toggleSorting(column.getIsSorted() === "asc")}
         >
           Nombres
-          <ArrowUpDown />
+          <ArrowUpDown className="ml-2 h-4 w-4" />
         </Button>
       ),
       cell: ({ row }) => (
@@ -108,11 +121,23 @@ export function LoansTable() {
     {
       accessorKey: "amount",
       header: "Monto",
-      cell: ({ row }) => {
-        return <div>S/. {row.original.amount}</div>;
-      },
+      cell: ({ row }) => <div>S/. {row.original.amount}</div>,
     },
-
+    {
+      accessorKey: "status",
+      header: "Estado",
+      cell: ({ row }) => (
+        <div
+          className={`text-white w-min px-2 text-center rounded-lg ${getLoanStatusColor(
+            row.original.status as keyof typeof LoanStatusColor
+          )}`}
+        >
+          {`${getLoanStatusText(
+            row.original.status as keyof typeof LoanStatusText
+          )}`}
+        </div>
+      ),
+    },
     {
       id: "actions",
       header: "Opciones",
@@ -121,33 +146,33 @@ export function LoansTable() {
           <DropdownMenuTrigger asChild>
             <Button variant="ghost" className="h-8 w-8 p-0">
               <span className="sr-only">Abrir menú</span>
-              <MoreHorizontal />
+              <MoreHorizontal className="h-4 w-4" />
             </Button>
           </DropdownMenuTrigger>
           <DropdownMenuContent align="end">
             <DropdownMenuItem
-              onClick={() => {
-                if (row.original.id) {
-                  console.log(row.original.id);
-                  fetchInstallments(row.original.id);
-                }
-              }}
+              onClick={() =>
+                row.original.id && fetchInstallments(row.original.id)
+              }
             >
-              <BookText />
+              <BookText className="mr-2 h-4 w-4" />
               Cuotas
-            </DropdownMenuItem>
-            <DropdownMenuSeparator />
-            <DropdownMenuItem>
-              <Edit />
-              Editar
             </DropdownMenuItem>
             <DropdownMenuSeparator />
             <DropdownMenuItem
               onClick={() => {
-                if (row.original.id) handleDelete(row.original.id);
+                setSelectedLoanId(row.original.id || null);
+                setOpenForm(true);
               }}
             >
-              <Trash />
+              <Edit className="mr-2 h-4 w-4" />
+              Editar
+            </DropdownMenuItem>
+            <DropdownMenuSeparator />
+            <DropdownMenuItem
+              onClick={() => row.original.id && handleDelete(row.original.id)}
+            >
+              <Trash className="mr-2 h-4 w-4" />
               Eliminar
             </DropdownMenuItem>
           </DropdownMenuContent>
@@ -155,6 +180,7 @@ export function LoansTable() {
       ),
     },
   ];
+
   const table = useReactTable({
     data: loans,
     columns,
@@ -162,88 +188,98 @@ export function LoansTable() {
     getCoreRowModel: getCoreRowModel(),
     getPaginationRowModel: getPaginationRowModel(),
     getSortedRowModel: getSortedRowModel(),
-    state: {
-      sorting,
-    },
+    state: { sorting },
   });
 
   return (
-    <div className="w-full">
-      <LoansInstallments
-        installments={installments}
-        onOpenChange={() => setInstallments([])}
-        isOpen={installments.length > 0}
-      />
-      <div className="flex items-center py-4">
-        <Input
-          placeholder="Filtrar nombres..."
-          onChange={(event) =>
-            table.getColumn("user.name")?.setFilterValue(event.target.value)
-          }
-          className="max-w-sm"
+    <>
+      <div className="w-full">
+        <LoansInstallments
+          installments={installments}
+          onOpenChange={() => setInstallments([])}
+          isOpen={installments.length > 0}
         />
-      </div>
-      <div className="rounded-md border">
-        <Table>
-          <TableHeader>
-            {table.getHeaderGroups().map((headerGroup) => (
-              <TableRow key={headerGroup.id}>
-                {headerGroup.headers.map((header) => (
-                  <TableHead key={header.id}>
-                    {flexRender(
-                      header.column.columnDef.header,
-                      header.getContext()
-                    )}
-                  </TableHead>
-                ))}
-              </TableRow>
-            ))}
-          </TableHeader>
-          <TableBody>
-            {table.getRowModel().rows.length ? (
-              table.getRowModel().rows.map((row) => (
-                <TableRow key={row.id}>
-                  {row.getVisibleCells().map((cell) => (
-                    <TableCell key={cell.id}>
+        <div className="flex items-center py-4">
+          <Input
+            placeholder="Filtrar nombres..."
+            onChange={(event) =>
+              table.getColumn("user.name")?.setFilterValue(event.target.value)
+            }
+            className="max-w-sm"
+          />
+        </div>
+        <div className="rounded-md border">
+          <Table>
+            <TableHeader>
+              {table.getHeaderGroups().map((headerGroup) => (
+                <TableRow key={headerGroup.id}>
+                  {headerGroup.headers.map((header) => (
+                    <TableHead key={header.id}>
                       {flexRender(
-                        cell.column.columnDef.cell,
-                        cell.getContext()
+                        header.column.columnDef.header,
+                        header.getContext()
                       )}
-                    </TableCell>
+                    </TableHead>
                   ))}
                 </TableRow>
-              ))
-            ) : (
-              <TableRow>
-                <TableCell
-                  colSpan={columns.length}
-                  className="h-24 text-center"
-                >
-                  No se encontraron resultados.
-                </TableCell>
-              </TableRow>
-            )}
-          </TableBody>
-        </Table>
+              ))}
+            </TableHeader>
+            <TableBody>
+              {table.getRowModel().rows.length ? (
+                table.getRowModel().rows.map((row) => (
+                  <TableRow key={row.id}>
+                    {row.getVisibleCells().map((cell) => (
+                      <TableCell key={cell.id}>
+                        {flexRender(
+                          cell.column.columnDef.cell,
+                          cell.getContext()
+                        )}
+                      </TableCell>
+                    ))}
+                  </TableRow>
+                ))
+              ) : (
+                <TableRow>
+                  <TableCell
+                    colSpan={columns.length}
+                    className="h-24 text-center"
+                  >
+                    No se encontraron resultados.
+                  </TableCell>
+                </TableRow>
+              )}
+            </TableBody>
+          </Table>
+        </div>
+        <div className="flex items-center justify-end space-x-2 py-4">
+          <Button
+            variant="outline"
+            size="sm"
+            onClick={() => table.previousPage()}
+            disabled={!table.getCanPreviousPage()}
+          >
+            Anterior
+          </Button>
+          <Button
+            variant="outline"
+            size="sm"
+            onClick={() => table.nextPage()}
+            disabled={!table.getCanNextPage()}
+          >
+            Siguiente
+          </Button>
+        </div>
       </div>
-      <div className="flex items-center justify-end space-x-2 py-4">
-        <Button
-          variant="outline"
-          size="sm"
-          onClick={() => table.previousPage()}
-          disabled={!table.getCanPreviousPage()}
-        >
-          Anterior
-        </Button>
-        <Button
-          variant="outline"
-          size="sm"
-          onClick={() => table.nextPage()}
-          disabled={!table.getCanNextPage()}
-        >
-          Siguiente
-        </Button>
-      </div>
-    </div>
+      <DialogForm
+        open={openForm}
+        onOpenChange={setOpenForm}
+        isEdit={!!selectedLoanId}
+        disabledTrigger
+      >
+        <LoanForm
+          loan={loans.find((loan) => loan.id === selectedLoanId) || null}
+        />
+      </DialogForm>
+    </>
   );
 }
